add federation plugin hook
