+++ crates/apub/Cargo.toml
@@ -47,6 +47,10 @@ html2md = "0.2.14"
 html2text = "0.6.0"
 stringreader = "0.1.1"
 enum_delegate = "0.2.0"
+extism = { version = "1.2.0", features = [
+  "register-filesystem",
+], default-features = false }
+extism-convert = { version = "1.2.0", default-features = false }
 
 [dev-dependencies]
 serial_test = { workspace = true }
